refactor: text enum 생성해서 관리하도록 수정 (#91)
