+++ DietLog-MVVM/View/SignIn/SignInText.swift
@@ -1,0 +1,20 @@
+//
+//  SignInText.swift
+//  DietLog-MVVM
+//
+//  Created by Jooyeon Kang on 6/9/24.
+//
+
+import Foundation
+
+enum SignInText {
+    static let dietLog = "DIET LOG"
+    static let title = "안녕하세요"
+    static let subTitle = "회원 가입을 진행해 주세요~"
+    static let nickname = "닉네임"
+    static let email = "이메일"
+    static let password = "비밀번호"
+    static let sns = "간단하게 가입해 보세요"
+    static let welcom = "환영합니다"
+    static let nicknameEmptyError = "닉네임을 입력해 주세요"
+}
