+++ DietLog-MVVM/View/SignIn/SignInViewController.swift
@@ -9,16 +9,6 @@ import UIKit
 import SnapKit
 import RxSwift
 
-enum SignInText {
-    static let dietLog = "DIET LOG"
-    static let title = "안녕하세요"
-    static let subTitle = "회원 가입을 진행해 주세요~"
-    static let nickname = "닉네임"
-    static let email = "이메일"
-    static let password = "비밀번호"
-    static let sns = "간단하게 가입해 보세요"
-}
-
 class SignInViewController: BaseViewController {
     
     // MARK: - UI Componet
@@ -235,9 +225,9 @@ class SignInViewController: BaseViewController {
                 var message = ""
                 
                 if result {
-                    message = "환영합니다"
+                    message = SignInText.welcom
                 } else {
-                    message = "닉네임을 입력해 주세요"
+                    message = SignInText.nicknameEmptyError
                 }
                 
                 self?.showAlertWithOKButton(title: nil, message: message) {
